Fix HomomorphismStructureOnMorphisms of AlgebroidFromDataTables for CompilerForCAP
Background: CompilerForCAP has a rule
`List( list_of_lists[index], func )` =>
`List( list_of_lists, list -> List( list, func ) )[index]`.
This rule assumes that `list_of_lists` is "homogeneous" with regard to
`func`, that is, that `func` can be applied to any element of
`list_of_lists`.

When compiling AdditiveClosure( AlgebroidFromDataTables ), `s_alpha` and
`s_gamma` arise as matrix entries, hence we are in the situation of the
above rule. However, `func` contains `coeffs_alpha[l]` and
`coeffs_gamma[r]`. Thus, `func` cannot be applied to any `s_alpha` or
`s_gamma` but only to the `s_alpha` and `s_gamma` corresponding to
`coeffs_alpha` and `coeffs_gamma`. Thus, the rule above can lead to
out-of-bounds errors. Due to (generalized) loop fusion this did not show up
in the final compilation result, but this was more or less by chance.
Accessing `s_alpha` and `s_gamma` in `func` explicitly avoids this issue.